server/requirements.txt
aniso8601==9.0.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
nltk==3.6.5
oauthlib==3.1.1
pycodestyle==2.8.0
python-dotenv==0.19.2
pytz==2021.3
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
textblob==0.17.1
toml==0.10.2
tqdm==4.62.3
tweepy==4.4.0
urllib3==1.26.7
vaderSentiment==3.3.2
Werkzeug==2.0.2
